requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
anthropic>=0.7.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
pydantic>=2.5.3
pytest>=7.4.4 